added more popups

+++ database_management.py
@@ -10,7 +10,7 @@ class DataManager:
         self.db_cursor: sql3.Cursor
         self.database = database
         self.popup_module = PopupModule(
-            "Yes", "No", "Unos Mušterije", "Da li želite unijeti mušteriju?"
+            "Yes", "No", "", ""
         )
 
     def db_connect(self, database):
@@ -35,6 +35,8 @@ class DataManager:
         self.db_disconnect()
 
     def db_insert_customer(self, table: str, customer: list):
+        self.popup_module.set_title('Unos')
+        self.popup_module.set_question('Da li želite spremiti unesene podatke?')
         answer = self.popup_module.confirmation_dialog()
         if answer:
             self.db_connect(self.database)
@@ -107,19 +109,23 @@ class DataManager:
             table_widget.insertRow(index + 1)
 
     def delete_selected_customer(self, table: str, table_widget: QTableWidget):
-        if (
-            table_widget.selectedItems() != []
-            and table_widget.selectedItems() is not None
-        ):
-            selected_row = table_widget.selectedItems()
-            self.db_connect(self.database)
-            self.db_cursor = self.db_link.cursor()
-            self.db_cursor.execute(
-                f"DELETE FROM {table} WHERE id = ?",
-                (selected_row[4].text(),),
-            )
-            self.db_disconnect()
-            table_widget.removeRow(table_widget.currentRow())
+        self.popup_module.set_title('Brisanje')
+        self.popup_module.set_question('Da li želite obrisati mušteriju iz baze podataka? \n (Upozorenje: Ova radnja je nepovratna!)')
+        answer = self.popup_module.confirmation_dialog()
+        if answer:
+            if (
+                table_widget.selectedItems() != []
+                and table_widget.selectedItems() is not None
+            ):
+                selected_row = table_widget.selectedItems()
+                self.db_connect(self.database)
+                self.db_cursor = self.db_link.cursor()
+                self.db_cursor.execute(
+                    f"DELETE FROM {table} WHERE id = ?",
+                    (selected_row[4].text(),),
+                )
+                self.db_disconnect()
+                table_widget.removeRow(table_widget.currentRow())
             # self.populate_customer_table(table, table_widget, (), "")
 
     def clear_customer_table(self, table_widget: QTableWidget):
